backend/pyproject.toml
[project]
name = "iu-pos-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.9",
    "django-cleanup>=9.0.0",
    "django-cors-headers>=4.9.0",
    "django-protected-media>=1.0.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
]
